example user data added
